Start to add eventing dashboards. (#20)
This creates sections for "topics" and "subscriptions" that filter on
their respective prefixes.

The topic section shows both topic metrics and subscription metrics for
topics matching the prefix. The subscription section shows the
subscription metrics for subscriptions matching the prefix.

Building on this, the broker ingress now has a customized dashboard that
takes advantage of the topic metrics to put them front-and-center. Also
building on this, the event recorder now includes a section with
subscription metrics for each even type.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>