Add outbound HTTP dashboard (#28)
This based on the metrics we export in our metricking transport.

There are some difference with Cloud Run native metrics:
1. we can't add `resource.type` to our metrics. I work around by not
using `resource.type` in the filter here.
2. the cloud run native metrics has `service_name` being a
`resource.label`, which AFAIK we can't add to our Prometheus metrics.
when following the instruction
[here](https://cloud.google.com/run/docs/tutorials/custom-metrics-opentelemetry-sidecar)
the `service_name` is a `metric.label`. As a result, we will need to add
that filter here, unlike the other widgets that can inherit from the
Dashboard's.

For 2. we could also add a template variable at the Dashboard level that
use a `metric.label`, but will look a bit confusing as we have two drop
downs that look exactly the same. Since we are baking `var.service_name`
to the dashboard name anyway, I think we may as well drop the template
filter from there.

This is how this change would look like (if the metrics are there)


![image](https://github.com/chainguard-dev/terraform-cloudrun-glue/assets/138266/a14961b7-cf69-4055-8601-0a87e4d31309)

---------

Signed-off-by: Nghia Tran <tcnghia@gmail.com>