add bucket-events module (#166)
This module sets up a Pub/Sub subscription for GCS events coming from a
bucket, and translates them to CloudEvents that it sends to a configured
Broker.

So far I've only tested this with a regional GCS bucket and
single-region topic and trampoline service. For multi-regional buckets
we arbitrarily pick a region in that multi-region that must be part of
the regional-networks. This isn't perfect but it's a first draft.

Fixes #142

---------

Signed-off-by: Jason Hall <jason@chainguard.dev>