bots: initial commit (#248)
This is a work in progress prototype demonstrating how we can write an
army of cute lil bots to manage our repos for us.

The basic idea is that we can write bots that subscribe to GH events (PR
opened, commit pushed, workflow finished, etc.), and those bots can take
actions on the subject of the event (the PR, the workflow), as easily as
possible.

This adds a GH client wrapper that handles the OctoSTS token
request/revoke, and tries to simplify the set of actions the bots are
trying to do.

The other concept touched on in this PR is "blocking labels": bots can
add labels to block merging (e.g., `blocking/dnm`, `blocking/bincapz`,
etc.), and the blocker bot will be responsible for seeing those labels
and passing/failing a check run on the PR. That check can be made a
required check, so that the existence of the labels blocks merging. We
can simulate "ack-merging" by removing the labels, which is recorded in
the PR's action log.

The idea is that bots have a way to put blocks in place that are
intentionally easy to ack (and audited!), and that humans can manually
add back as needed (also audited). All of this in Go, a real
non-YAMLbash language.

The GH client wrapper also adds a method to create-or-update a comment
"owned" by the bot.

---------

Signed-off-by: Jason Hall <jason@chainguard.dev>