support ReaderAt for GH file retrievals (#339)
Create new `Fetch*` alternative functions for retrieving `io.ReaderAt`s
for the helper methods that deal with get'ing arbitrary files from GH
(logs and artifacts).

This stops us from arbitrarily buffering the entire contents into
memory, and restricts the buffer size to 1mb for now. This currently
isn't configurable, but wouldn't be hard to make it configurable if
needed.

This also marks the old `Get*` functions as deprecated.