github-bots handle token error (#358)
When the octo-sts policy was incorrect we ran into two errors.

The first one in `handleGithubResponse` was a cryptic segfault where
`resp` was nil so accessing `resp.Response` threw an error

The second was in the `Close()` function we check if `ts.tok` is nil
before trying to revoke. Previously it was never nil, it was an pointer
to a Token struct with an empty token. This caused us to fail on token
revocation

Signed-off-by: James Petersen <jpetersenames@gmail.com>