github-bots: remove trigger from the name to allow more custom chars
I tried creating a bot with a brief descriptive name, for example my-team-update-check  but got an error during the apply:

```
(""bot-trigger-my-team-update-check-bj0e") doesn't match regexp "^[a-z](?:[-a-z0-9]{4,28}[a-z0-9])$"
```

`bot-trigger-` is always added as a prefix, which makes it pretty easy to go over the 28 char regex restriction.

Could we remove the `trigger-` part?

Signed-off-by: James Rawlings <jrawlings@chainguard.dev>