enable cron module to run non-Go images (#734)
Currently the cron module will perform a `ko` build atop a given base
image. This PR expands it to make `importpath` optional, in which case
we just run the base image itself.

This is useful in case we want to use `cron` module with `apko`, to
customize the command or the `run-as` user.

---------

Signed-off-by: Nghia Tran <tcnghia@gmail.com>