github-bots/sdk: add function to set comments on GitHub Issues
This also moves the main functionality of the existing setcomments for a PR, into a helper which is reused by both the set comments for PR and Issue.

Issues don't always have the repository set on the main object, so also have a helper function to get the owner + repo name from the boject if set, else fall back to getting it from the repository URL.

Added test cases for this too.

Signed-off-by: rawlingsj <jrawlings@chainguard.dev>