github-events: Add merged extension for merged pull requests (#800)
Add a new field to CloudEvents indicating when a pull request is merged.

This adds a boolean "merged" extension to events when a pull request has
been merged, making it easier for downstream consumers to filter and
process merged PRs without having to parse the event payload.

A filter to consume only merged PRs might look like this:

```
      filter = {
        "type"   = "dev.chainguard.github.pull_request"
        "action" = "closed"
        "merged" = "true"
      }
```

Signed-off-by: James Rawlings <jrawlings@chainguard.dev>