panic_filter: match text regex 'runtime/panic' (#798)
I propose a change to this alert to also handle runtime panics as part
of the condition that fires this alert. I am not so sure we want to fire
this alert even if the severity is NOT at the ERROR level. But I'd
consider awkward to log panics as just info/warn log entries.

related to:
https://chainguard-dev.slack.com/archives/C05CM0DM7RV/p1745920565303639

---------

Signed-off-by: hectorj2f <hector@chainguard.dev>
Signed-off-by: Hector Fernandez <hector@chainguard.dev>
Co-authored-by: Jason Hall <jason@chainguard.dev>