fix: gke - set default of pod_ipv4_cidr_block to null to prevent invalid empty string
Change the default value for the `pod_ipv4_cidr_block` variable in `modules/gke/variables.tf` from `""` (empty string) to `null`.

When the default is set to `""` (empty string), any configuration value of `null` gets converted to an empty string by Terraform because of `optional(string, "")`. The Google provider rejects an empty string for this field, expecting either a valid CIDR block or no value at all (`null`). By setting the default to `null`, we allow the variable to be truly unset if not specified, which aligns with provider expectations and prevents errors.

Signed-off-by: Maxime Gréau <maxime.greau@chainguard.dev>