add input variable `additional-service-account-accessors` to `module/secret` (#875)
There are cases where more than one service is required to access a
secret value (e.g. when using a redis module with auth, any viewers will
also need access to the auth string to make the connection).

Retain the `_binding` resource in the secret module, but allow
additional accessors to be defined.

---------

Signed-off-by: Colin Douglas <colin@chainguard.dev>