workqueue: Add error status details to not reenqueue errors. (#874)
Sometimes we don't want to re-enqueue and error because it's not
retriable. This is something that's going to be very client specific,
since we don't know under what conditions particular error codes may or
may not be retried.

This change allows clients to attach additional details to the error,
allowing them to inform the dispatcher when certain errors should not be
reenqueued. This preservese the existing behavior where keys are
reenqueued on error by default.