cloud-events-recorder: Fix behavior around retrying `writer.Close` (#896)
Currently, we're only retrying the respective `writer.Close` operation
up to 3 times. However, retrying `Close` itself doesn't seem to be doing
anything useful and just returns the same error over and over again.

Instead, this retries the entire operation. The a successful write
operation consists of writing the respective blob **and** successfully
calling `Close`. So in addition to the fixed retrying, this also moves
deleting the respective files to **after** calling `Close` successfully,
to ensure we're not losing state in error cases.