Truncate WIF provider pool display name (#916)
Sometimes people want to pass in longer names.

With this change, if the `name` passed to the `github-wif-provider` is longer than 23 characters, it will be truncated when it's referred to in the pool display name.

This has come up in some internal PRs:

- https://github.com/chainguard-dev/infra/pull/433
- https://github.com/chainguard-dev/infra/pull/436
- https://github.com/chainguard-dev/infra/pull/712

I think another good approach would be to validate the `name` variable,
restricting it to 23 characters, and to provide a further `description`
variable so that users can provide the full name without truncation.