httpmetrics: Backfill transport tests and fix to actually check the labels
Curried metrics don't seem to filter in testing at all, meaning that the labels can be whatever and the tests still pass. That's no bueno.

Instead, this asserts that the labels are all actually as expected.

It also backfills tests for the in-flight metric to ensure we're getting the metrics that we want.