workqueue: Fix dashboards to be per-location as applicable (#970)
Some of our aggregations in the workqueue dashboard are problematic
today in that they skew metrics because we're running in multiple
regions but load is heavily skewed to one of them. MEAN aggregations
across all locations thus underestimate the actual load on the system.

This changes all tiles to be broken out by location to avoid such
skewing and to allow us to see what's really going on across the
regions.