Create `regional-go-reconciler` and a `reconciler` dashboard.
This creates a new streamlined mechanism for creating reconcilers.

The `reconciler` dashboard makes it easy to create a single dashboard for the workqueue pieces as well as the reconciler service itself.  It also supports legacy services (e.g. syncer) that were stood up with separate workqueue/syncer modules, but may want to take advantage of this.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>