Create `regional-go-reconciler` and a `reconciler` dashboard. (#1089)
This creates a new streamlined mechanism for creating reconcilers.

The `reconciler` dashboard makes it easy to create a single dashboard
for the workqueue pieces as well as the reconciler service itself. It
also supports legacy services (e.g. syncer) that were stood up with
separate workqueue/syncer modules, but may want to take advantage of
this.

Based on:
https://github.com/chainguard-dev/terraform-infra-common/pull/1088

---------

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>
Co-authored-by: octo-sts[bot] <157150467+octo-sts[bot]@users.noreply.github.com>