Drop regional scope, randomize global bucket name.
This eliminates the ability to specify a regional scope for workqueue, regional-go-reconciler, and the associated dashboards.

This also shifts the (now) solve GCS bucket to use a randomized suffix to reduce the possibility of collisions given the global namespace of GCS buckets.

This change is technically lossy as it deletes the live GCS bucket backing the workqueue and replaces it.

However, the nature of workqueues is that you can "resync" (e.g. backstop) and recover from any lost data (the K8s equivalent is a controller restart, this is why controllers resync at startup).

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>