Add a "status manager" for the github reconciler.
The status manager is designed to enable us to manage status in a way that is closer to what we see in kubernetes for the PR context.

Treating the commit SHA as the "generation" and associating status with the generation we are observing.

In principle this would let us observe an older generation's status, but I am skipping any lookback to be conscious of API quota (if we want this, then perhaps a dedicated method for fetching all of the observed generations' Statuses can be added later).

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>