add a module to concat two strings, subject to a given maximum length (#1130)
There are a few resources that have constraints on the length of their
names. We currently apply the same name to a bunch of subresources
within modules, and if one violates a maximum length requirement the
entire module fails. This module intends to allow us to only shorten
names when necessary, and otherwise continue using the caller provided
names as is.

Signed-off-by: Colin Douglas <colin@chainguard.dev>