Create a helper token source that requeues (#1135)
The requeues are to avoid spinning during quota exhaustion events.

Note that the reconciler (when it directly) hits the quota exhaustion,
already uses RequeueAfter with when the quota resets.

This aims to detect a class of these where that reset is less clear, and
avoid having us spin on the key and DLQ.

---------

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>
Co-authored-by: octo-sts[bot] <157150467+octo-sts[bot]@users.noreply.github.com>