Add a "path" github resource. (#1158)
Plumb through some preliminary support for "path" type github resources,
where we use the blob or tree URL as the key.

This adds support for these in the StatusManager as well, but updates
how we use check run names a bit to shard check run names by path, since
I could be reconciling N > 1 paths for a particular SHA, and so we
should have a status for each key.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>