Add a github path reconciler module.
This module basically sets up all of the scaffolding to reconcile files within a github repository.

There are two main triggers:
1. A periodic cron (every X hours)
2. A push event listener (immediate)

When these run, they enumerate either all of the files in the repo, or all of the changed files in the push.  We apply the file patterns to each of them, and when they match, we take the substring that the pattern captures as the key.

The intent of this is to be able to match e.g. `images/foo` when `images/foo/bar/baz.blah` changes by having a pattern like:
```
(images/[^/]+)/.*
```

There are a couple funky interactions between this and some of the current workqueue semantics, but I am going to fix that in a parallel change.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>