Change `workqueue` to prefer lowest timestamp.
Currently we choose the highest timestamp, which inhibits our ability to force trying a key earlier.

If we have a low priority key queued with a long delay, and we want to try it with high priority now, the result is a high priority with the long delay timestamp.

In https://github.com/chainguard-dev/terraform-infra-common/pull/1159 I want to be able to periodically resync N keys by loading the key space and queuing all of the keys in the future, but this may be hours in the future.

When a key actually changes, we want to be able to process it immediately with high priority, but the current semantics mean that if there is an outstanding resync that some keys will get starved until the resync queued keys are processed.

```
ok      github.com/chainguard-dev/terraform-infra-common/pkg/workqueue/gcs      316.861s
```

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>