Make pullrequesturl lowest-common-denominator.
We only get this for branches on the upstream, but not from forks.

Unless/Until this information is populated, this is an unreliable trigger for reconciliation.

It is also notable that most reconcilers don't actually react to check run completions, and where we do we want to handle forks, so we have had to work around this anyways.

So including this creates both unnecessary churn in the reconcilers while not being sufficient for the cases that actually care.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>