Switch to a higher file-limit approach.
Apparently CompareCommits "only" returns up to 300 files and 250 commits, so this is an unreliably way of enumerating changes.

The new approach uses `GetTree` which returns all of the files and their SHAs with a limit of 100k files, so we can use this to determine file changes with only a couple API calls.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>