fix(github): Actually apply secondary rate limiting
I was trying to write tests for
https://github.com/chainguard-dev/terraform-infra-common/pull/1210
but kept hitting strange issues with the transport not being changed,
despite me writing code that looked exactly like
`WithSecondaryRateLimitWaiter`.

I'm pretty sure the way we override the transport in the GitHub client is
wrong.

This changes it to make a whole new GitHub client.

It might be reasonable to propose a `WithClient` API to upstream.

To be extra sure, we could roll out
https://github.com/chainguard-dev/terraform-infra-common/pull/1208 first,
observe a lack of metrics, and then merge this.