workqueue: use GCP-optimized logger (#1218)
The line being removed overrides the work done by underscore-importing
`clog/gcp/init` and forces in a slog TextHandler instead.

This is probably not intentional, but if it is I'd love to understand
why.

The result of the TextHandler is that GCP logs don't include log
attributes, and those lines aren't correlated with other lines belonging
to the same trace, if there are any.

Signed-off-by: Jason Hall <jason@chainguard.dev>