Try adding a batch size. (#1222)
With the shift to multi-regional buckets, we see load balancing across
regions, but I believe that when container concurrency is low large
batches queuing across different regions can cause us to keep more
instances alive than we probably should have.

e.g. for a workqueue with concurrency of 10 across 3 regions and a
reconciler with a container concurrency of 1, I would expect us to
typically have 3-4 instances per region under constant load. In
practice, we see closer to 10 per region.

My working theory is that the reason this is so high is that we are
frequently seeing a single pass of the dispatcher launch 10 concurrent
requests, keeping all 10 instances in a region alive (and this behavior
bounces between regions enough to keep ~30 instances alive).

See the slack thread here:
https://chainguard-dev.slack.com/archives/C090L6K0Q07/p1763387732630259

---------

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>
Co-authored-by: octo-sts[bot] <157150467+octo-sts[bot]@users.noreply.github.com>