Increase the max jitter percentage to 100 (#1257)
It would be nice to make the max jitter inversely proportional to `d`,
but I'm struggling to imagine how to do that without knowing the maximum
possible back-off time from GitHub to use as the constant to compare `d`
against.

In the mean time, I think we should increase the max jitter percentage
to 100%. This is because, from what I can see, the backoff time is
usually in the seconds to minutes level of magnitude and 10% of these
ranges are not usually sufficient to smooth out the spike of retries
that get queued up.

---------

Signed-off-by: Joe Borg <joe.borg@chainguard.dev>